+++ st2common/st2common/runners/python_action_wrapper.py
@@ -16,6 +16,7 @@
 from __future__ import absolute_import
 import os
 import sys
+from six.moves import input
 
 # Note: This work-around is required to fix the issue with other Python modules which live
 # inside this directory polluting and masking sys.path for Python runner actions.
@@ -251,6 +252,8 @@ if __name__ == '__main__':
                         help='Pack config serialized as JSON')
     parser.add_argument('--parameters', required=False,
                         help='Serialized action parameters')
+    parser.add_argument('--stdin-parameters', required=False, action='store_true',
+                        help='Serialized action parameters via stdin')
     parser.add_argument('--user', required=False,
                         help='User who triggered the action execution')
     parser.add_argument('--parent-args', required=False,
@@ -261,11 +264,25 @@ if __name__ == '__main__':
     args = parser.parse_args()
 
     config = json.loads(args.config) if args.config else {}
-    parameters = args.parameters
-    parameters = json.loads(parameters) if parameters else {}
     user = args.user
     parent_args = json.loads(args.parent_args) if args.parent_args else []
     log_level = args.log_level
+
+    parameters = {}
+
+    if args.parameters:
+        LOG.debug('Getting parameters from argument')
+        args_parameters = args.parameters
+        args_parameters = json.loads(args_parameters) if args_parameters else {}
+        parameters.update(args_parameters)
+
+    if args.stdin_parameters:
+        LOG.debug('Getting parameters from stdin')
+        stdin_parameters = json.loads(input())
+        stdin_parameters = stdin_parameters.get('parameters', {})
+        parameters.update(stdin_parameters)
+
+    LOG.debug('Received parameters: %s', parameters)
 
     assert isinstance(parent_args, list)
     obj = PythonActionWrapper(pack=args.pack,